V2.11 a full code for test meal compara

+++ STM32_Demo/utilities/include/bsp.h
@@ -131,8 +131,8 @@ extern uint8_t   CoinFlag;			 //用于退币机的标志
 extern uint8_t   TemperatureCur;
 extern uint8_t   Current ;    //主状态
 extern uint8_t   CurrentPointer; //餐品出餐状态
-extern uint16_t  erro_flag;  //错误状态
-extern int32_t   erro_record;  //错误标记位
+extern uint32_t  erro_flag;  //错误状态
+extern uint32_t   erro_record;  //错误标记位
 extern uint16_t MoneyPayBack_Already; //上传用的已退币数量值
 extern uint16_t MoneyPayBack_Already_total; //计算总的退币数
 extern bool MoneyBackCnt_Already;
@@ -149,7 +149,7 @@ void StateSend(void);
 void DataUpload(char takemeal_flag);
 void PrintTickFun(unsigned char *PrintTickFlag);
 void PowerupAbnormalHandle(int32_t erro_record);
-void AbnormalHandle(uint16_t erro);
+void AbnormalHandle(uint32_t erro);
 void PollAbnormalHandle(void);
 void SaveUserData(void);
 void ReadUserData(void);
